Update resend API
新增
1. 提供重新上拋特定效益的API
2. 提供重新上拋特定時間區間內的特定效益的API
更新
1. 支援多種效益在同一Job下執行API

+++ src/api/server.ts
@@ -30,11 +30,42 @@ export class Server {
    * API Server
    */
   private readonly server = express();
+  /**
+   * 效益計算Job
+   */
+  private benefit = new Map<string, BnftTemplate>();
+  /**
+   * 單一實例
+   */
+  public static instance = new Server();
+
+  private constructor() {
+    // 啟動API Server
+    this.start();
+  }
 
   /**
-   * @param benefit 效益計算Job
+   * 取得效益計算Job清單
+   *
+   * @method private
+   * @return 回傳效益計算Job清單
    */
-  constructor(private benefit: BnftTemplate) {}
+  private getBenefitList(): BnftTemplate[] {
+    return Array.from(this.benefit).map((benefit) => benefit[1]);
+  }
+
+  /**
+   * 取得效益Key名稱
+   *
+   * @method private
+   * @param req 請求
+   * @return 回傳效益Key名稱
+   */
+  private getBenefitKey(req: express.Request): string {
+    const systemId = req.query.systemId;
+    const typeId = req.query.typeId;
+    return `${systemId}.${typeId}`;
+  }
 
   /**
    * 重新上拋所有效益
@@ -49,8 +80,35 @@ export class Server {
   ): Promise<void> {
     const timestamp = new Date(parseInt(req.query.timestamp as string, 10));
     timestamp.setDate(timestamp.getDate() + 1);
-    const result = await this.benefit.execute(timestamp).toPromise();
+    const result = await Promise.all(
+      this.getBenefitList().map((benefit) =>
+        benefit.execute(timestamp).toPromise()
+      )
+    );
     res.send(result);
+  }
+
+  /**
+   * 重新上拋特定效益
+   *
+   * @method private
+   * @param req 請求
+   * @param res 回應
+   */
+  private async sendSpecificBenefit(
+    req: express.Request,
+    res: express.Response
+  ): Promise<void> {
+    const key = this.getBenefitKey(req);
+    const timestamp = new Date(parseInt(req.query.timestamp as string, 10));
+    timestamp.setDate(timestamp.getDate() + 1);
+    const benefit = this.benefit.get(key);
+    if (benefit) {
+      const result = await benefit.execute(timestamp).toPromise();
+      res.send(result);
+    } else {
+      res.send(null);
+    }
   }
 
   /**
@@ -70,8 +128,43 @@ export class Server {
         req.body.map(async (date) => {
           const timestamp = new Date(parseInt(date, 10));
           timestamp.setDate(timestamp.getDate() + 1);
-          const result = await this.benefit.execute(timestamp).toPromise();
-          resultList.push(result);
+          const result = await Promise.all(
+            this.getBenefitList().map((benefit) =>
+              benefit.execute(timestamp).toPromise()
+            )
+          );
+          resultList.push(...result);
+        })
+      );
+      res.send(resultList);
+    } else {
+      res.send('body format error');
+    }
+  }
+
+  /**
+   * 重新上拋特定時間區間內的特定效益
+   *
+   * @method private
+   * @param req 請求
+   * @param res 會應
+   */
+  public async sendSpecificBenefitByTimeGroup(
+    req: express.Request,
+    res: express.Response
+  ): Promise<void> {
+    if (Array.isArray(req.body) && req.body.length > 0) {
+      const key = this.getBenefitKey(req);
+      const resultList: ProducePayloadModel<Bnft.BenefitSaving>[] = [];
+      await Promise.all(
+        req.body.map(async (date) => {
+          const timestamp = new Date(parseInt(date, 10));
+          timestamp.setDate(timestamp.getDate() + 1);
+          const benefit = await this.benefit.get(key);
+          if (benefit) {
+            const result = await benefit.execute(timestamp).toPromise();
+            resultList.push(result);
+          }
         })
       );
       res.send(resultList);
@@ -96,8 +189,10 @@ export class Server {
       files.map((file) => {
         const benefit = fs.readFileSync(`backup/${file}`, { encoding: 'utf8' });
         try {
-          this.benefit.send(JSON.parse(benefit));
-          fs.unlinkSync(`backup/${file}`);
+          if (this.getBenefitList().length > 0) {
+            this.getBenefitList()[0].send(JSON.parse(benefit));
+            fs.unlinkSync(`backup/${file}`);
+          }
         } catch (error) {
           this.logger.warn('Send backup benefit failed');
           this.logger.warn(benefit);
@@ -111,21 +206,40 @@ export class Server {
   /**
    * 啟動API Server
    *
-   * @method public
-   * @param port API端口
+   * @method private
    * @return 回傳物件本身
    */
-  public start(port = 3000): Server {
+  private start(): Server {
+    const port = 3000;
     this.server.use(bodyParser.json());
     this.server.post('/benefit/send', this.sendBenefit.bind(this));
     this.server.post(
+      '/specific/benefit/send',
+      this.sendSpecificBenefit.bind(this)
+    );
+    this.server.post(
       '/benefit/send/timestamps',
       this.sendBenefitByTimeGroup.bind(this)
+    );
+    this.server.post(
+      '/specific/benefit/send/timestamps',
+      this.sendSpecificBenefitByTimeGroup.bind(this)
     );
     this.server.post('/benefit/backup', this.sendBackupBenefit.bind(this));
     this.server.use('/', swaggerUi.serve, swaggerUi.setup(SWAGGER_DOC));
     this.logger.info(`Listen api port ${port}`);
     this.server.listen(port);
     return this;
+  }
+
+  /**
+   * 註冊效益計算Job
+   *
+   * @method public
+   * @param benefit 效益計算Job
+   */
+  public register(benefit: BnftTemplate): void {
+    const key = `${benefit.config.systemId}.${benefit.config.typeId}`;
+    this.benefit.set(key, benefit);
   }
 }
